+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/mysql/repository/AttendanceRepository.java
@@ -1,6 +1,6 @@
-package com.blockpage.memberservice.adaptor.infrastructure.repository;
+package com.blockpage.memberservice.adaptor.infrastructure.mysql.repository;
 
-import com.blockpage.memberservice.adaptor.infrastructure.entity.AttendanceEntity;
+import com.blockpage.memberservice.adaptor.infrastructure.mysql.entity.AttendanceEntity;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
